+++ apps/web/src/app/_landing/footer.tsx
@@ -1,5 +1,3 @@
-"use client";
-
 import Link from "next/link";
 import type { Route } from "next";
